+++ src/entities/meal/index.js
@@ -1,2 +1,3 @@
 export { MealList } from './ui/meal-list.jsx';
 export { MealCard } from './ui/meal-card.jsx';
+export { useMeals } from './model/use-meals.js';
