+++ src/entities/meal/ui/meal-card.jsx
