+++ src/entities/meal/ui/meal-card.module.css
